+++ frost-core/src/signature.rs
@@ -1,5 +1,6 @@
 //! Schnorr signatures over prime order groups (or subgroups)
 
+use alloc::vec::Vec;
 use debugless_unwrap::DebuglessUnwrap;
 
 use crate::{Ciphersuite, Element, Error, Field, Group, Scalar};
@@ -72,7 +73,7 @@ where
 
     /// Converts this signature to its [`Ciphersuite::SignatureSerialization`] in bytes.
     pub fn serialize(&self) -> C::SignatureSerialization {
-        let mut bytes = vec![];
+        let mut bytes = Vec::<u8>::new();
 
         bytes.extend(<C::Group>::serialize(&self.R).as_ref());
         bytes.extend(<<C::Group as Group>::Field>::serialize(&self.z).as_ref());
@@ -117,8 +118,8 @@ where
     }
 }
 
-impl<C: Ciphersuite> std::fmt::Debug for Signature<C> {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+impl<C: Ciphersuite> core::fmt::Debug for Signature<C> {
+    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
         f.debug_struct("Signature")
             .field("R", &hex::encode(<C::Group>::serialize(&self.R).as_ref()))
             .field(