+++ frost-ed448/src/lib.rs
@@ -5,6 +5,8 @@
 #![doc = include_str!("../README.md")]
 #![doc = document_features::document_features!()]
 
+extern crate alloc;
+
 use std::collections::BTreeMap;
 
 use ed448_goldilocks::{